+++ packages/relayer/src/index.ts
@@ -30,8 +30,7 @@ export interface Relayer {
 }
 
 export { LocalRelayer } from './local-relayer'
-export { RpcRelayer } from './rpc-relayer'
-export { ProviderRelayer } from './provider-relayer'
+export { RpcRelayer, proto as RpcRelayerProto } from './rpc-relayer'
 
 export function isRelayer(cand: any): cand is Relayer {
   return cand && cand.estimateGasLimits !== undefined && cand.gasRefundOptions !== undefined &&