+++ packages/api/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @0xsequence/api
 
+## 0.16.1
+
+### Patch Changes
+
+- api: add legacy types for bw compat
+
+## 0.16.0
+
+### Minor Changes
+
+- relayer as its own service separate from chaind
+
+## 0.15.1
+
+### Patch Changes
+
+- update api clients
+
 ## 0.15.0
 
 ### Patch Changes