+++ packages/chaind/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @0xsequence/chaind
 
+## 0.16.0
+
+### Minor Changes
+
+- relayer as its own service separate from chaind
+
+## 0.15.1
+
+### Patch Changes
+
+- update api clients
+
 ## 0.15.0
 
 ### Patch Changes