+++ packages/multicall/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @0xsequence/multicall
 
+## 0.16.0
+
+### Minor Changes
+
+- relayer as its own service separate from chaind
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @0xsequence/abi@0.16.0
+  - @0xsequence/network@0.16.0
+  - @0xsequence/utils@0.16.0
+
+## 0.15.1
+
+### Patch Changes
+
+- update api clients
+- Updated dependencies [undefined]
+  - @0xsequence/abi@0.15.1
+  - @0xsequence/network@0.15.1
+  - @0xsequence/utils@0.15.1
+
 ## 0.14.3
 
 ### Patch Changes