+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @0xsequence/provider
 
+## 0.16.1
+
+### Patch Changes
+
+- @0xsequence/auth@0.16.1
+
+## 0.16.0
+
+### Minor Changes
+
+- relayer as its own service separate from chaind
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @0xsequence/abi@0.16.0
+  - @0xsequence/auth@0.16.0
+  - @0xsequence/config@0.16.0
+  - @0xsequence/network@0.16.0
+  - @0xsequence/transactions@0.16.0
+  - @0xsequence/utils@0.16.0
+  - @0xsequence/wallet@0.16.0
+
+## 0.15.1
+
+### Patch Changes
+
+- update api clients
+- Updated dependencies [undefined]
+  - @0xsequence/abi@0.15.1
+  - @0xsequence/auth@0.15.1
+  - @0xsequence/config@0.15.1
+  - @0xsequence/network@0.15.1
+  - @0xsequence/transactions@0.15.1
+  - @0xsequence/utils@0.15.1
+  - @0xsequence/wallet@0.15.1
+
 ## 0.15.0
 
 ### Patch Changes