+++ packages/relayer/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @0xsequence/relayer
 
+## 0.16.0
+
+### Minor Changes
+
+- relayer as its own service separate from chaind
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @0xsequence/abi@0.16.0
+  - @0xsequence/chaind@0.16.0
+  - @0xsequence/config@0.16.0
+  - @0xsequence/transactions@0.16.0
+  - @0xsequence/utils@0.16.0
+
+## 0.15.1
+
+### Patch Changes
+
+- update api clients
+- Updated dependencies [undefined]
+  - @0xsequence/abi@0.15.1
+  - @0xsequence/chaind@0.15.1
+  - @0xsequence/config@0.15.1
+  - @0xsequence/transactions@0.15.1
+  - @0xsequence/utils@0.15.1
+
 ## 0.15.0
 
 ### Minor Changes